outbound-calling-speech-assistant-openai-realtime-api-python/requirements.txt
fastapi
uvicorn
twilio
websockets
sqlmodel
psycopg2-binary
chromadb
python-dotenv
google-genai
numpy
scipy
